ТЗ9. Итеррация1. Гитхаб
