+++ src/management/DurationTimeAdapter.java
@@ -1,0 +1,21 @@
+package management;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class DurationTimeAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final Duration duration) throws IOException {
+        jsonWriter.value(duration.toMinutes());
+    }
+
+    @Override
+    public Duration read(final JsonReader jsonReader) throws IOException {
+        return Duration.parse(jsonReader.nextString());
+    }
+}
