+++ src/management/LocalDateTimeAdapter.java
@@ -1,0 +1,23 @@
+package management;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+    private static final DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy-MM-dd.HH:mm:ss");
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final LocalDateTime localDate) throws IOException {
+        jsonWriter.value(localDate.format(dtf));
+    }
+
+    @Override
+    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
+        return LocalDateTime.parse(jsonReader.nextString(), dtf);
+    }
+}
